Fixed audio
Currently, every time you click right on your mouse the audio is played. Call the audio by the group name. you can add multiple sounds in one group and the code will randomly choose one of those groups. This is so that the sound has some distinction each time it is called.